% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template_available.R
\name{activity_available}
\alias{activity_available}
\title{Show available activities}
\usage{
activity_available(show_devel = FALSE)
}
\arguments{
\item{show_devel}{(Logical) Show activities currently under development, Default: FALSE}
}
\value{
Character string of the activities available
}
\description{
Show available activities
}
