% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activity_download.R
\name{activity_download}
\alias{activity_download}
\title{Download activity}
\usage{
activity_download(path = "./Scripts", activity_type = "poppk",
  overwrite = FALSE, devel = FALSE)
}
\arguments{
\item{path}{path to copy to, defaults to a 'Scripts' folder}

\item{activity_type}{type of activity, default poppk}

\item{overwrite}{PARAM_DESCRIPTION, Default: FALSE}

\item{devel}{(Logical) Get activities currently under development, Default: FALSE}
}
\value{
A set of template script files for the specified activity
}
\description{
Downloads a set of template scripts for pmx activities
}
