% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_report_readme.R
\name{get_report_readme}
\alias{get_report_readme}
\title{get the report readme docx file}
\usage{
get_report_readme(path = ".", proj_type = "poppk_workflow")
}
\arguments{
\item{path}{path to copy to, defaults to current working dir}

\item{proj_type}{project template type, default poppk_workflow}
}
\description{
get the report readme docx file
}

