% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{list_files}
\alias{list_files}
\alias{list_files_and_dirs}
\title{list files, including hidden ones, in a directory}
\usage{
list_files(path = ".", all.files = TRUE, recursive = TRUE, no.. = TRUE,
  ..., normalize = FALSE, must_work = TRUE)

list_files_and_dirs(...)
}
\arguments{
\item{path}{a character vector of full path names; the default corresponds to the working directory, Default: '.'}

\item{all.files}{a logical value. If FALSE, only the names of visible files are returned.
If TRUE, all file names will be returned., Default: TRUE}

\item{recursive}{logical. Should the listing recurse into directories?, Default: TRUE}

\item{no..}{logical. Should both "." and ".." be excluded also from non-recursive listings?, Default: TRUE}

\item{...}{args to pass to [dir()]}

\item{normalize}{whether to convert file paths to canonical form for the platform, Default: FALSE}

\item{must_work}{logical: if TRUE then an error is given if the result cannot be determined; if NA then a warning., Default: TRUE}

\item{...}{parameters to pass to list_files}
}
\value{
vector of file paths

a list containing files and dirs elements
}
\description{
list files, including hidden ones, in a directory

list files, including hidden ones, in a directory
}
