% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{package_filepath}
\alias{package_filepath}
\title{Find package file path}
\usage{
package_filepath(...)
}
\arguments{
\item{...}{parameters to pass to system.file}
}
\value{
file path
}
\description{
A light wrapper around system.file to get extract file or dir from package
}
