% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_dir.R
\name{source_dir}
\alias{source_dir}
\title{Source all scripts in dir}
\usage{
source_dir(path, trace = TRUE, ...)
}
\arguments{
\item{path}{Path to directory of files to source, Default: '.'}

\item{trace}{, Default: TRUE}

\item{...}{other arguments to pass to source}
}
\value{
Prints files souced to console
}
\description{
Sources all .R/.r scripts in a directory
}
