% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template_download.R
\name{template_download}
\alias{template_download}
\title{Download project templates}
\usage{
template_download(path = "./Scripts", template_type = "poppk_workflow",
  overwrite = FALSE)
}
\arguments{
\item{path}{path to copy to, defaults to a 'Scripts' folder}

\item{template_type}{type of project template, default poppk_workflow}
}
\value{
set of template R and Rmarkdown script files
}
\description{
Downloads a set of template scripts for nlme projects in nonmem
}
