% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_report.R
\name{init_project}
\alias{init_project}
\title{initialize a project}
\usage{
init_project(path = ".", proj_type = "poppk_workflow")
}
\arguments{
\item{path}{path to copy to, defaults to current working dir}

\item{proj_type}{project template type, default poppk_workflow}
}
\description{
initialize a project
}
