% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{package_filepath}
\alias{package_filepath}
\title{get a file or dir from a the package}
\usage{
package_filepath(...)
}
\arguments{
\item{...}{parameters to pass to system.file}
}
\description{
get a file or dir from a the package
}
\details{
a light wrapper around system.file
}
