% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{source_dir}
\alias{source_dir}
\title{Source all .R/.r scripts in a directory}
\usage{
source_dir(path, trace = TRUE, ...)
}
\arguments{
\item{path}{file paths}

\item{trace}{}

\item{...}{arguments to pass to source}
}
\description{
Source all .R/.r scripts in a directory
}
