% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{source_dir}
\alias{source_dir}
\title{source all .R files in a directory}
\usage{
source_dir(path, trace = TRUE, ...)
}
\arguments{
\item{path}{directory paths}

\item{trace}{Default = TRUE}
}
\description{
source all .R files in a directory
}
