% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template_available.R
\name{activity_available}
\alias{activity_available}
\title{Show available project templates}
\usage{
activity_available()
}
\value{
set of template R and Rmarkdown script files
}
\description{
Show available project templates
}
