% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template_available.R
\name{template_available}
\alias{template_available}
\title{Show available project templates}
\usage{
template_available()
}
\value{
set of template R and Rmarkdown script files
}
\description{
Show available project templates
}
